Mieraf exception and validation

+++ Application/Features/Post/Commands/DeletePost/DeletePostCommandHandler.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Application.Exceptions;
 using AutoMapper;
 using MediatR;
 
@@ -19,6 +20,10 @@ public class DeletePostCommandHandler : IRequestHandler<DeletePostCommand,Unit>
 
     public async Task<Unit> Handle(DeletePostCommand command, CancellationToken cancellationToken)
     {
+        bool res = await _postRepository.Exists(command.PostId);
+        if(res == false)
+             throw new NotFoundException($"Post with id {command.PostId} does't exist!", command);
+             
         var post =  await _postRepository.DeleteAsync(command.PostId);
         return Unit.Value;
     }
